smbios: Move table build tools into an include file.
This will let things in other files (like IPMI) build SMBIOS tables.
